refactor(tests): Move tests of `load` to their own module
