+++ skore/tests/unit/project/test_project.py
@@ -1,4 +1,3 @@
-import os
 from io import BytesIO
 
 import altair
@@ -16,11 +15,7 @@ from skore.exceptions import (
     InvalidProjectNameError,
     ProjectCreationError,
 )
-from skore.project import (
-    Project,
-    create,
-    load,
-)
+from skore.project import create
 from skore.project.create import _validate_project_name
 from skore.view.view import View
 
@@ -143,33 +138,6 @@ def test_put_rf_model(in_memory_project, monkeypatch):
     model.fit(numpy.array([[1, 2], [3, 4]]), [0, 1])
     in_memory_project.put("rf_model", model)  # ScikitLearnModelItem
     assert isinstance(in_memory_project.get("rf_model"), RandomForestClassifier)
-
-
-def test_load_no_project():
-    with pytest.raises(FileNotFoundError):
-        load("/empty")
-
-
-@pytest.fixture
-def fake_project_path(tmp_path):
-    """Create a project at `tmp_path` and return its absolute path."""
-    # Project path must end with ".skore"
-    project_path = tmp_path.parent / (tmp_path.name + ".skore")
-    os.mkdir(project_path)
-    os.mkdir(project_path / "items")
-    os.mkdir(project_path / "views")
-    return project_path
-
-
-def test_load_absolute_path(fake_project_path):
-    p = load(fake_project_path)
-    assert isinstance(p, Project)
-
-
-def test_load_relative_path(fake_project_path):
-    os.chdir(fake_project_path.parent)
-    p = load(fake_project_path.name)
-    assert isinstance(p, Project)
 
 
 def test_put(in_memory_project):
